add: add modularity for history operations
